+++ drape_frontend/route_renderer.cpp
@@ -52,7 +52,7 @@ std::vector<float> const kPreviewPointRadiusInPixel =
 int const kArrowAppearingZoomLevel = 14;
 int const kInvalidGroup = -1;
 
-size_t const kPreviewPointsCount = 512;
+uint32_t const kPreviewPointsCount = 512;
 double const kPreviewAnimationSpeed = 3.0;
 double const kPreviewAnimationScale = 0.3;
